Slope bias function
We have generated a dataset by varying the true slope, true intercept, sample size and error variance in x and y to calculate the bias in the slope. We then wrote the bias function in terms of the factors mentioned that were seen to be significantly related to the bias value.